split out ConfigBuilder
