+++ benches/bench.rs
@@ -56,7 +56,7 @@ fn benchmark_simple_pipe(c: &mut Criterion) {
         Benchmark::new("pipe", move |b| {
             b.iter_with_setup(
                 || {
-                    let mut config = config::Config::empty();
+                    let mut config = config::Config::builder();
                     config.add_source(
                         "in",
                         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -72,7 +72,7 @@ fn benchmark_simple_pipe(c: &mut Criterion) {
                     let mut rt = runtime();
                     let (output_lines, topology) = rt.block_on(async move {
                         let output_lines = CountReceiver::receive_lines(out_addr);
-                        let (topology, _crash) = start_topology(config, false).await;
+                        let (topology, _crash) = start_topology(config.build(), false).await;
                         wait_for_tcp(in_addr).await;
                         (output_lines, topology)
                     });
@@ -107,7 +107,7 @@ fn benchmark_simple_pipe_with_tiny_lines(c: &mut Criterion) {
         Benchmark::new("pipe_with_tiny_lines", move |b| {
             b.iter_with_setup(
                 || {
-                    let mut config = config::Config::empty();
+                    let mut config = config::Config::builder();
                     config.add_source(
                         "in",
                         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -123,7 +123,7 @@ fn benchmark_simple_pipe_with_tiny_lines(c: &mut Criterion) {
                     let mut rt = runtime();
                     let (output_lines, topology) = rt.block_on(async move {
                         let output_lines = CountReceiver::receive_lines(out_addr);
-                        let (topology, _crash) = start_topology(config, false).await;
+                        let (topology, _crash) = start_topology(config.build(), false).await;
                         wait_for_tcp(in_addr).await;
                         (output_lines, topology)
                     });
@@ -158,7 +158,7 @@ fn benchmark_simple_pipe_with_huge_lines(c: &mut Criterion) {
         Benchmark::new("pipe_with_huge_lines", move |b| {
             b.iter_with_setup(
                 || {
-                    let mut config = config::Config::empty();
+                    let mut config = config::Config::builder();
                     config.add_source(
                         "in",
                         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -174,7 +174,7 @@ fn benchmark_simple_pipe_with_huge_lines(c: &mut Criterion) {
                     let mut rt = runtime();
                     let (output_lines, topology) = rt.block_on(async move {
                         let output_lines = CountReceiver::receive_lines(out_addr);
-                        let (topology, _crash) = start_topology(config, false).await;
+                        let (topology, _crash) = start_topology(config.build(), false).await;
                         wait_for_tcp(in_addr).await;
                         (output_lines, topology)
                     });
@@ -210,7 +210,7 @@ fn benchmark_simple_pipe_with_many_writers(c: &mut Criterion) {
         Benchmark::new("pipe_with_many_writers", move |b| {
             b.iter_with_setup(
                 || {
-                    let mut config = config::Config::empty();
+                    let mut config = config::Config::builder();
                     config.add_source(
                         "in",
                         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -226,7 +226,7 @@ fn benchmark_simple_pipe_with_many_writers(c: &mut Criterion) {
                     let mut rt = runtime();
                     let (output_lines, topology) = rt.block_on(async move {
                         let output_lines = CountReceiver::receive_lines(out_addr);
-                        let (topology, _crash) = start_topology(config, false).await;
+                        let (topology, _crash) = start_topology(config.build(), false).await;
                         wait_for_tcp(in_addr).await;
                         (output_lines, topology)
                     });
@@ -274,7 +274,7 @@ fn benchmark_interconnected(c: &mut Criterion) {
         Benchmark::new("interconnected", move |b| {
             b.iter_with_setup(
                 || {
-                    let mut config = config::Config::empty();
+                    let mut config = config::Config::builder();
                     config.add_source(
                         "in1",
                         sources::socket::SocketConfig::make_tcp_config(in_addr1),
@@ -302,7 +302,7 @@ fn benchmark_interconnected(c: &mut Criterion) {
                     let (output_lines1, output_lines2, topology) = rt.block_on(async move {
                         let output_lines1 = CountReceiver::receive_lines(out_addr1);
                         let output_lines2 = CountReceiver::receive_lines(out_addr2);
-                        let (topology, _crash) = start_topology(config, false).await;
+                        let (topology, _crash) = start_topology(config.build(), false).await;
                         wait_for_tcp(in_addr1).await;
                         wait_for_tcp(in_addr2).await;
                         (output_lines1, output_lines2, topology)
@@ -341,7 +341,7 @@ fn benchmark_transforms(c: &mut Criterion) {
         Benchmark::new("transforms", move |b| {
             b.iter_with_setup(
                 || {
-                    let mut config = config::Config::empty();
+                    let mut config = config::Config::builder();
                     config.add_source(
                         "in",
                         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -374,7 +374,7 @@ fn benchmark_transforms(c: &mut Criterion) {
                     let mut rt = runtime();
                     let (output_lines, topology) = rt.block_on(async move {
                         let output_lines = CountReceiver::receive_lines(out_addr);
-                        let (topology, _crash) = start_topology(config, false).await;
+                        let (topology, _crash) = start_topology(config.build(), false).await;
                         wait_for_tcp(in_addr).await;
                         (output_lines, topology)
                     });
@@ -453,7 +453,7 @@ fn benchmark_complex(c: &mut Criterion) {
         Benchmark::new("complex", move |b| {
             b.iter_with_setup(
                 || {
-                    let mut config = config::Config::empty();
+                    let mut config = config::Config::builder();
                     config.add_source(
                         "in1",
                         sources::socket::SocketConfig::make_tcp_config(in_addr1),
@@ -552,7 +552,7 @@ fn benchmark_complex(c: &mut Criterion) {
                         let output_lines_sampled = CountReceiver::receive_lines(out_addr_sampled);
                         let output_lines_200 = CountReceiver::receive_lines(out_addr_200);
                         let output_lines_404 = CountReceiver::receive_lines(out_addr_404);
-                        let (topology, _crash) = start_topology(config, false).await;
+                        let (topology, _crash) = start_topology(config.build(), false).await;
                         wait_for_tcp(in_addr1).await;
                         wait_for_tcp(in_addr2).await;
                         (
