+++ benches/files.rs
@@ -42,7 +42,7 @@ fn benchmark_files_without_partitions(c: &mut Criterion) {
                 source.include.push(input.clone());
                 source.data_dir = Some(data_dir);
 
-                let mut config = config::Config::empty();
+                let mut config = config::Config::builder();
                 config.add_source("in", source);
                 config.add_sink(
                     "out",
@@ -57,7 +57,7 @@ fn benchmark_files_without_partitions(c: &mut Criterion) {
 
                 let mut rt = runtime();
                 let (topology, input) = rt.block_on(async move {
-                    let (topology, _crash) = start_topology(config, false).await;
+                    let (topology, _crash) = start_topology(config.build(), false).await;
 
                     let mut options = OpenOptions::new();
                     options.create(true).write(true);
