+++ src/config/builder.rs
@@ -1,0 +1,158 @@
+use super::{
+    default_data_dir, Config, GlobalOptions, SinkConfig, SinkOuter, SourceConfig, TestDefinition,
+    TransformConfig, TransformOuter,
+};
+use crate::event;
+use indexmap::IndexMap;
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Debug, Default)]
+#[serde(deny_unknown_fields)]
+pub struct ConfigBuilder {
+    #[serde(flatten)]
+    pub global: GlobalOptions,
+    #[serde(default)]
+    pub sources: IndexMap<String, Box<dyn SourceConfig>>,
+    #[serde(default)]
+    pub sinks: IndexMap<String, SinkOuter>,
+    #[serde(default)]
+    pub transforms: IndexMap<String, TransformOuter>,
+    #[serde(default)]
+    pub tests: Vec<TestDefinition>,
+}
+
+impl Clone for ConfigBuilder {
+    fn clone(&self) -> Self {
+        // This is a hack around the issue of cloning
+        // trait objects. So instead to clone the config
+        // we first serialize it into JSON, then back from
+        // JSON. Originally we used TOML here but TOML does not
+        // support serializing `None`.
+        let json = serde_json::to_vec(self).unwrap();
+        serde_json::from_slice(&json[..]).unwrap()
+    }
+}
+
+impl ConfigBuilder {
+    // TODO: compilation here
+    pub fn build(self) -> Config {
+        Config {
+            global: self.global,
+            sources: self.sources,
+            sinks: self.sinks,
+            transforms: self.transforms,
+            tests: self.tests,
+            expansions: Default::default(),
+        }
+    }
+
+    pub fn add_source<S: SourceConfig + 'static>(&mut self, name: &str, source: S) {
+        self.sources.insert(name.to_string(), Box::new(source));
+    }
+
+    pub fn add_sink<S: SinkConfig + 'static>(&mut self, name: &str, inputs: &[&str], sink: S) {
+        let inputs = inputs.iter().map(|&s| s.to_owned()).collect::<Vec<_>>();
+        let sink = SinkOuter {
+            buffer: Default::default(),
+            healthcheck: true,
+            inner: Box::new(sink),
+            inputs,
+        };
+
+        self.sinks.insert(name.to_string(), sink);
+    }
+
+    pub fn add_transform<T: TransformConfig + 'static>(
+        &mut self,
+        name: &str,
+        inputs: &[&str],
+        transform: T,
+    ) {
+        let inputs = inputs.iter().map(|&s| s.to_owned()).collect::<Vec<_>>();
+        let transform = TransformOuter {
+            inner: Box::new(transform),
+            inputs,
+        };
+
+        self.transforms.insert(name.to_string(), transform);
+    }
+
+    pub fn append(&mut self, with: Self) -> Result<(), Vec<String>> {
+        let mut errors = Vec::new();
+
+        if self.global.data_dir.is_none() || self.global.data_dir == default_data_dir() {
+            self.global.data_dir = with.global.data_dir;
+        } else if with.global.data_dir != default_data_dir()
+            && self.global.data_dir != with.global.data_dir
+        {
+            // If two configs both set 'data_dir' and have conflicting values
+            // we consider this an error.
+            errors.push("conflicting values for 'data_dir' found".to_owned());
+        }
+
+        // If the user has multiple config files, we must *merge* log schemas until we meet a
+        // conflict, then we are allowed to error.
+        let default_schema = event::LogSchema::default();
+        if with.global.log_schema != default_schema {
+            // If the set value is the default, override it. If it's already overridden, error.
+            if self.global.log_schema.host_key() != default_schema.host_key()
+                && self.global.log_schema.host_key() != with.global.log_schema.host_key()
+            {
+                errors.push("conflicting values for 'log_schema.host_key' found".to_owned());
+            } else {
+                self.global
+                    .log_schema
+                    .set_host_key(with.global.log_schema.host_key().clone());
+            }
+            if self.global.log_schema.message_key() != default_schema.message_key()
+                && self.global.log_schema.message_key() != with.global.log_schema.message_key()
+            {
+                errors.push("conflicting values for 'log_schema.message_key' found".to_owned());
+            } else {
+                self.global
+                    .log_schema
+                    .set_message_key(with.global.log_schema.message_key().clone());
+            }
+            if self.global.log_schema.timestamp_key() != default_schema.timestamp_key()
+                && self.global.log_schema.timestamp_key() != with.global.log_schema.timestamp_key()
+            {
+                errors.push("conflicting values for 'log_schema.timestamp_key' found".to_owned());
+            } else {
+                self.global
+                    .log_schema
+                    .set_timestamp_key(with.global.log_schema.timestamp_key().clone());
+            }
+        }
+
+        with.sources.keys().for_each(|k| {
+            if self.sources.contains_key(k) {
+                errors.push(format!("duplicate source name found: {}", k));
+            }
+        });
+        with.sinks.keys().for_each(|k| {
+            if self.sinks.contains_key(k) {
+                errors.push(format!("duplicate sink name found: {}", k));
+            }
+        });
+        with.transforms.keys().for_each(|k| {
+            if self.transforms.contains_key(k) {
+                errors.push(format!("duplicate transform name found: {}", k));
+            }
+        });
+        with.tests.iter().for_each(|wt| {
+            if self.tests.iter().any(|t| t.name == wt.name) {
+                errors.push(format!("duplicate test name found: {}", wt.name));
+            }
+        });
+        if !errors.is_empty() {
+            return Err(errors);
+        }
+
+        self.sources.extend(with.sources);
+        self.sinks.extend(with.sinks);
+        self.transforms.extend(with.transforms);
+        self.tests.extend(with.tests);
+
+        Ok(())
+    }
+}
