+++ src/config/mod.rs
@@ -13,6 +13,7 @@ use snafu::{ResultExt, Snafu};
 use std::fs::DirBuilder;
 use std::path::PathBuf;
 
+mod builder;
 pub mod component;
 mod diff;
 mod loading;
@@ -24,20 +25,13 @@ pub use diff::ConfigDiff;
 pub use loading::{load_from_paths, load_from_str, process_paths, CONFIG_PATHS};
 pub use validation::check;
 
-#[derive(Deserialize, Serialize, Debug)]
-#[serde(deny_unknown_fields)]
+#[derive(Debug, Default)]
 pub struct Config {
-    #[serde(flatten)]
     pub global: GlobalOptions,
-    #[serde(default)]
     pub sources: IndexMap<String, Box<dyn SourceConfig>>,
-    #[serde(default)]
     pub sinks: IndexMap<String, SinkOuter>,
-    #[serde(default)]
     pub transforms: IndexMap<String, TransformOuter>,
-    #[serde(default)]
     pub tests: Vec<TestDefinition>,
-    #[serde(skip)]
     expansions: IndexMap<String, Vec<String>>,
 }
 
@@ -319,51 +313,9 @@ pub enum TestCondition {
     String(String),
 }
 
-// Helper methods for programming construction during tests
 impl Config {
-    pub fn empty() -> Self {
-        Self {
-            global: GlobalOptions {
-                data_dir: None,
-                log_schema: event::LogSchema::default(),
-            },
-            sources: IndexMap::new(),
-            sinks: IndexMap::new(),
-            transforms: IndexMap::new(),
-            tests: Vec::new(),
-            expansions: IndexMap::new(),
-        }
-    }
-
-    pub fn add_source<S: SourceConfig + 'static>(&mut self, name: &str, source: S) {
-        self.sources.insert(name.to_string(), Box::new(source));
-    }
-
-    pub fn add_sink<S: SinkConfig + 'static>(&mut self, name: &str, inputs: &[&str], sink: S) {
-        let inputs = inputs.iter().map(|&s| s.to_owned()).collect::<Vec<_>>();
-        let sink = SinkOuter {
-            buffer: Default::default(),
-            healthcheck: true,
-            inner: Box::new(sink),
-            inputs,
-        };
-
-        self.sinks.insert(name.to_string(), sink);
-    }
-
-    pub fn add_transform<T: TransformConfig + 'static>(
-        &mut self,
-        name: &str,
-        inputs: &[&str],
-        transform: T,
-    ) {
-        let inputs = inputs.iter().map(|&s| s.to_owned()).collect::<Vec<_>>();
-        let transform = TransformOuter {
-            inner: Box::new(transform),
-            inputs,
-        };
-
-        self.transforms.insert(name.to_string(), transform);
+    pub fn builder() -> builder::ConfigBuilder {
+        Default::default()
     }
 
     /// Expand a logical component name (i.e. from the config file) into the names of the
@@ -418,100 +370,8 @@ impl Config {
             Ok(())
         }
     }
-
-    pub fn append(&mut self, with: Self) -> Result<(), Vec<String>> {
-        let mut errors = Vec::new();
-
-        if self.global.data_dir.is_none() || self.global.data_dir == default_data_dir() {
-            self.global.data_dir = with.global.data_dir;
-        } else if with.global.data_dir != default_data_dir()
-            && self.global.data_dir != with.global.data_dir
-        {
-            // If two configs both set 'data_dir' and have conflicting values
-            // we consider this an error.
-            errors.push("conflicting values for 'data_dir' found".to_owned());
-        }
-
-        // If the user has multiple config files, we must *merge* log schemas until we meet a
-        // conflict, then we are allowed to error.
-        let default_schema = event::LogSchema::default();
-        if with.global.log_schema != default_schema {
-            // If the set value is the default, override it. If it's already overridden, error.
-            if self.global.log_schema.host_key() != default_schema.host_key()
-                && self.global.log_schema.host_key() != with.global.log_schema.host_key()
-            {
-                errors.push("conflicting values for 'log_schema.host_key' found".to_owned());
-            } else {
-                self.global
-                    .log_schema
-                    .set_host_key(with.global.log_schema.host_key().clone());
-            }
-            if self.global.log_schema.message_key() != default_schema.message_key()
-                && self.global.log_schema.message_key() != with.global.log_schema.message_key()
-            {
-                errors.push("conflicting values for 'log_schema.message_key' found".to_owned());
-            } else {
-                self.global
-                    .log_schema
-                    .set_message_key(with.global.log_schema.message_key().clone());
-            }
-            if self.global.log_schema.timestamp_key() != default_schema.timestamp_key()
-                && self.global.log_schema.timestamp_key() != with.global.log_schema.timestamp_key()
-            {
-                errors.push("conflicting values for 'log_schema.timestamp_key' found".to_owned());
-            } else {
-                self.global
-                    .log_schema
-                    .set_timestamp_key(with.global.log_schema.timestamp_key().clone());
-            }
-        }
-
-        with.sources.keys().for_each(|k| {
-            if self.sources.contains_key(k) {
-                errors.push(format!("duplicate source name found: {}", k));
-            }
-        });
-        with.sinks.keys().for_each(|k| {
-            if self.sinks.contains_key(k) {
-                errors.push(format!("duplicate sink name found: {}", k));
-            }
-        });
-        with.transforms.keys().for_each(|k| {
-            if self.transforms.contains_key(k) {
-                errors.push(format!("duplicate transform name found: {}", k));
-            }
-        });
-        with.tests.iter().for_each(|wt| {
-            if self.tests.iter().any(|t| t.name == wt.name) {
-                errors.push(format!("duplicate test name found: {}", wt.name));
-            }
-        });
-        if !errors.is_empty() {
-            return Err(errors);
-        }
-
-        self.sources.extend(with.sources);
-        self.sinks.extend(with.sinks);
-        self.transforms.extend(with.transforms);
-        self.tests.extend(with.tests);
-
-        Ok(())
-    }
-
     pub fn typecheck(&self) -> Result<(), Vec<String>> {
         validation::typecheck(self)
-    }
-}
-
-impl Clone for Config {
-    fn clone(&self) -> Self {
-        // This is a hack around the issue of cloning
-        // trait objects. So instead to clone the config
-        // we first serialize it into JSON, then back from
-        // JSON. Originally we used TOML here but TOML does not
-        // support serializing `None`.
-        let json = serde_json::to_vec(self).unwrap();
-        serde_json::from_slice(&json[..]).unwrap()
     }
 }
 
@@ -526,12 +386,12 @@ fn healthcheck_default() -> bool {
     feature = "transforms-json_parser"
 ))]
 mod test {
-    use super::Config;
+    use super::{builder::ConfigBuilder, load_from_str};
     use std::path::PathBuf;
 
     #[test]
     fn default_data_dir() {
-        let config: Config = toml::from_str(
+        let config = load_from_str(
             r#"
       [sources.in]
       type = "file"
@@ -553,7 +413,7 @@ mod test {
 
     #[test]
     fn default_schema() {
-        let config: Config = toml::from_str(
+        let config = load_from_str(
             r#"
       [sources.in]
       type = "file"
@@ -580,7 +440,7 @@ mod test {
 
     #[test]
     fn custom_schema() {
-        let config: Config = toml::from_str(
+        let config = load_from_str(
             r#"
       [log_schema]
       host_key = "this"
@@ -606,7 +466,7 @@ mod test {
 
     #[test]
     fn config_append() {
-        let mut config: Config = toml::from_str(
+        let mut config: ConfigBuilder = toml::from_str(
             r#"
       [sources.in]
       type = "file"
@@ -657,7 +517,7 @@ mod test {
 
     #[test]
     fn config_append_collisions() {
-        let mut config: Config = toml::from_str(
+        let mut config: ConfigBuilder = toml::from_str(
             r#"
       [sources.in]
       type = "file"
