+++ src/sources/statsd/mod.rs
@@ -128,7 +128,7 @@ mod test {
         let in_addr = next_addr();
         let out_addr = next_addr();
 
-        let mut config = config::Config::empty();
+        let mut config = config::Config::builder();
         config.add_source("in", StatsdConfig { address: in_addr });
         config.add_sink(
             "out",
@@ -141,7 +141,7 @@ mod test {
             },
         );
 
-        let (topology, _crash) = start_topology(config, false).await;
+        let (topology, _crash) = start_topology(config.build(), false).await;
 
         let bind_addr = next_addr();
         let socket = std::net::UdpSocket::bind(&bind_addr).unwrap();
