+++ src/topology/mod.rs
@@ -51,7 +51,7 @@ pub async fn start_validated(
     let mut running_topology = RunningTopology {
         inputs: HashMap::new(),
         outputs: HashMap::new(),
-        config: Config::empty(),
+        config: Config::default(),
         shutdown_coordinator: SourceShutdownCoordinator::default(),
         source_tasks: HashMap::new(),
         tasks: HashMap::new(),
@@ -637,7 +637,7 @@ mod tests {
 
     #[tokio::test]
     async fn topology_doesnt_reload_new_data_dir() {
-        let mut old_config = Config::empty();
+        let mut old_config = Config::builder();
         old_config.add_source("in", SocketConfig::make_tcp_config(next_addr()));
         old_config.add_sink(
             "out",
@@ -650,12 +650,12 @@ mod tests {
         old_config.global.data_dir = Some(Path::new("/asdf").to_path_buf());
         let mut new_config = old_config.clone();
 
-        let (mut topology, _crash) = start_topology(old_config, false).await;
+        let (mut topology, _crash) = start_topology(old_config.build(), false).await;
 
         new_config.global.data_dir = Some(Path::new("/qwerty").to_path_buf());
 
         topology
-            .reload_config_and_respawn(new_config, false)
+            .reload_config_and_respawn(new_config.build(), false)
             .await
             .unwrap();
 
@@ -680,7 +680,7 @@ mod reload_tests {
     async fn topology_reuse_old_port() {
         let address = next_addr();
 
-        let mut old_config = Config::empty();
+        let mut old_config = Config::builder();
         old_config.add_source("in1", SplunkConfig::on(address));
         old_config.add_sink(
             "out",
@@ -691,7 +691,7 @@ mod reload_tests {
             },
         );
 
-        let mut new_config = Config::empty();
+        let mut new_config = Config::builder();
         new_config.add_source("in2", SplunkConfig::on(address));
         new_config.add_sink(
             "out",
@@ -702,9 +702,9 @@ mod reload_tests {
             },
         );
 
-        let (mut topology, _crash) = start_topology(old_config, false).await;
+        let (mut topology, _crash) = start_topology(old_config.build(), false).await;
         assert!(topology
-            .reload_config_and_respawn(new_config, false)
+            .reload_config_and_respawn(new_config.build(), false)
             .await
             .unwrap());
     }
@@ -713,7 +713,7 @@ mod reload_tests {
     async fn topology_rebuild_old() {
         let address = next_addr();
 
-        let mut old_config = Config::empty();
+        let mut old_config = Config::builder();
         old_config.add_source("in1", SplunkConfig::on(address));
         old_config.add_sink(
             "out",
@@ -724,7 +724,7 @@ mod reload_tests {
             },
         );
 
-        let mut new_config = Config::empty();
+        let mut new_config = Config::builder();
         old_config.add_source("in1", SplunkConfig::on(address));
         new_config.add_source("in2", SplunkConfig::on(address));
         new_config.add_sink(
@@ -736,9 +736,9 @@ mod reload_tests {
             },
         );
 
-        let (mut topology, _crash) = start_topology(old_config, false).await;
+        let (mut topology, _crash) = start_topology(old_config.build(), false).await;
         assert!(!topology
-            .reload_config_and_respawn(new_config, false)
+            .reload_config_and_respawn(new_config.build(), false)
             .await
             .unwrap());
     }
@@ -747,7 +747,7 @@ mod reload_tests {
     async fn topology_old() {
         let address = next_addr();
 
-        let mut old_config = Config::empty();
+        let mut old_config = Config::builder();
         old_config.add_source("in1", SplunkConfig::on(address));
         old_config.add_sink(
             "out",
@@ -758,9 +758,9 @@ mod reload_tests {
             },
         );
 
-        let (mut topology, _crash) = start_topology(old_config.clone(), false).await;
+        let (mut topology, _crash) = start_topology(old_config.clone().build(), false).await;
         assert!(topology
-            .reload_config_and_respawn(old_config, false)
+            .reload_config_and_respawn(old_config.build(), false)
             .await
             .unwrap());
     }
@@ -779,7 +779,7 @@ mod source_finished_tests {
 
     #[tokio::test]
     async fn sources_finished() {
-        let mut old_config = Config::empty();
+        let mut old_config = Config::builder();
         let generator = GeneratorConfig::repeat(vec!["text".to_owned()], 1, None);
         old_config.add_source("in", generator);
         old_config.add_sink(
@@ -791,7 +791,7 @@ mod source_finished_tests {
             },
         );
 
-        let (topology, _crash) = start_topology(old_config, false).await;
+        let (topology, _crash) = start_topology(old_config.build(), false).await;
 
         timeout(Duration::from_secs(2), topology.sources_finished().compat())
             .await
@@ -868,7 +868,7 @@ mod transient_state_tests {
 
     #[tokio::test]
     async fn closed_source() {
-        let mut old_config = Config::empty();
+        let mut old_config = Config::builder();
         let (trigger_old, source) = MockSourceConfig::new();
         old_config.add_source("in", source);
         old_config.add_transform(
@@ -882,7 +882,7 @@ mod transient_state_tests {
         old_config.add_sink("out1", &["trans"], BlackholeConfig { print_amount: 1000 });
         old_config.add_sink("out2", &["trans"], BlackholeConfig { print_amount: 1000 });
 
-        let mut new_config = Config::empty();
+        let mut new_config = Config::builder();
         let (_trigger_new, source) = MockSourceConfig::new();
         new_config.add_source("in", source);
         new_config.add_transform(
@@ -895,7 +895,7 @@ mod transient_state_tests {
         );
         new_config.add_sink("out1", &["trans"], BlackholeConfig { print_amount: 1000 });
 
-        let (mut topology, _crash) = start_topology(old_config, false).await;
+        let (mut topology, _crash) = start_topology(old_config.build(), false).await;
 
         trigger_old.cancel();
 
@@ -903,7 +903,7 @@ mod transient_state_tests {
         finished.compat().await.unwrap();
 
         assert!(topology
-            .reload_config_and_respawn(new_config, false)
+            .reload_config_and_respawn(new_config.build(), false)
             .await
             .unwrap());
     }
@@ -912,7 +912,7 @@ mod transient_state_tests {
     async fn remove_sink() {
         trace_init();
 
-        let mut old_config = Config::empty();
+        let mut old_config = Config::builder();
         old_config.add_source("in", StdinConfig::default());
         old_config.add_transform(
             "trans",
@@ -925,7 +925,7 @@ mod transient_state_tests {
         old_config.add_sink("out1", &["trans"], BlackholeConfig { print_amount: 1000 });
         old_config.add_sink("out2", &["trans"], BlackholeConfig { print_amount: 1000 });
 
-        let mut new_config = Config::empty();
+        let mut new_config = Config::builder();
         new_config.add_source("in", StdinConfig::default());
         new_config.add_transform(
             "trans",
@@ -937,9 +937,9 @@ mod transient_state_tests {
         );
         new_config.add_sink("out1", &["trans"], BlackholeConfig { print_amount: 1000 });
 
-        let (mut topology, _crash) = start_topology(old_config, false).await;
+        let (mut topology, _crash) = start_topology(old_config.build(), false).await;
         assert!(topology
-            .reload_config_and_respawn(new_config, false)
+            .reload_config_and_respawn(new_config.build(), false)
             .await
             .unwrap());
     }
@@ -948,7 +948,7 @@ mod transient_state_tests {
     async fn remove_transform() {
         trace_init();
 
-        let mut old_config = Config::empty();
+        let mut old_config = Config::builder();
         old_config.add_source("in", StdinConfig::default());
         old_config.add_transform(
             "trans1",
@@ -969,7 +969,7 @@ mod transient_state_tests {
         old_config.add_sink("out1", &["trans1"], BlackholeConfig { print_amount: 1000 });
         old_config.add_sink("out2", &["trans2"], BlackholeConfig { print_amount: 1000 });
 
-        let mut new_config = Config::empty();
+        let mut new_config = Config::builder();
         new_config.add_source("in", StdinConfig::default());
         new_config.add_transform(
             "trans1",
@@ -981,9 +981,9 @@ mod transient_state_tests {
         );
         new_config.add_sink("out1", &["trans1"], BlackholeConfig { print_amount: 1000 });
 
-        let (mut topology, _crash) = start_topology(old_config, false).await;
+        let (mut topology, _crash) = start_topology(old_config.build(), false).await;
         assert!(topology
-            .reload_config_and_respawn(new_config, false)
+            .reload_config_and_respawn(new_config.build(), false)
             .await
             .unwrap());
     }
