+++ tests/syslog.rs
@@ -27,7 +27,7 @@ async fn test_tcp_syslog() {
     let in_addr = next_addr();
     let out_addr = next_addr();
 
-    let mut config = config::Config::empty();
+    let mut config = config::Config::builder();
     config.add_source(
         "in",
         SyslogConfig::new(Mode::Tcp {
@@ -39,7 +39,7 @@ async fn test_tcp_syslog() {
 
     let output_lines = CountReceiver::receive_lines(out_addr);
 
-    let (topology, _crash) = start_topology(config, false).await;
+    let (topology, _crash) = start_topology(config.build(), false).await;
     // Wait for server to accept traffic
     wait_for_tcp(in_addr).await;
 
@@ -81,7 +81,7 @@ async fn test_unix_stream_syslog() {
     let in_path = tempfile::tempdir().unwrap().into_path().join("stream_test");
     let out_addr = next_addr();
 
-    let mut config = config::Config::empty();
+    let mut config = config::Config::builder();
     config.add_source(
         "in",
         SyslogConfig::new(Mode::Unix {
@@ -92,7 +92,7 @@ async fn test_unix_stream_syslog() {
 
     let output_lines = CountReceiver::receive_lines(out_addr);
 
-    let (topology, _crash) = start_topology(config, false).await;
+    let (topology, _crash) = start_topology(config.build(), false).await;
 
     // Wait for server to accept traffic
     while std::os::unix::net::UnixStream::connect(&in_path).is_err() {
@@ -143,7 +143,7 @@ async fn test_octet_counting_syslog() {
     let in_addr = next_addr();
     let out_addr = next_addr();
 
-    let mut config = config::Config::empty();
+    let mut config = config::Config::builder();
     config.add_source(
         "in",
         SyslogConfig::new(Mode::Tcp {
@@ -155,7 +155,7 @@ async fn test_octet_counting_syslog() {
 
     let output_lines = CountReceiver::receive_lines(out_addr);
 
-    let (topology, _crash) = start_topology(config, false).await;
+    let (topology, _crash) = start_topology(config.build(), false).await;
     // Wait for server to accept traffic
     wait_for_tcp(in_addr).await;
