+++ tests/tcp.rs
@@ -23,7 +23,7 @@ async fn pipe() {
     let in_addr = next_addr();
     let out_addr = next_addr();
 
-    let mut config = config::Config::empty();
+    let mut config = config::Config::builder();
     config.add_source(
         "in",
         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -36,7 +36,7 @@ async fn pipe() {
 
     let mut output_lines = CountReceiver::receive_lines(out_addr);
 
-    let (topology, _crash) = start_topology(config, false).await;
+    let (topology, _crash) = start_topology(config.build(), false).await;
     // Wait for server to accept traffic
     wait_for_tcp(in_addr).await;
 
@@ -61,7 +61,7 @@ async fn sample() {
     let in_addr = next_addr();
     let out_addr = next_addr();
 
-    let mut config = config::Config::empty();
+    let mut config = config::Config::builder();
     config.add_source(
         "in",
         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -83,7 +83,7 @@ async fn sample() {
 
     let mut output_lines = CountReceiver::receive_lines(out_addr);
 
-    let (topology, _crash) = start_topology(config, false).await;
+    let (topology, _crash) = start_topology(config.build(), false).await;
     // Wait for server to accept traffic
     wait_for_tcp(in_addr).await;
 
@@ -118,7 +118,7 @@ async fn fork() {
     let out_addr1 = next_addr();
     let out_addr2 = next_addr();
 
-    let mut config = config::Config::empty();
+    let mut config = config::Config::builder();
     config.add_source(
         "in",
         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -137,7 +137,7 @@ async fn fork() {
     let mut output_lines1 = CountReceiver::receive_lines(out_addr1);
     let mut output_lines2 = CountReceiver::receive_lines(out_addr2);
 
-    let (topology, _crash) = start_topology(config, false).await;
+    let (topology, _crash) = start_topology(config.build(), false).await;
     // Wait for server to accept traffic
     wait_for_tcp(in_addr).await;
 
@@ -172,7 +172,7 @@ async fn merge_and_fork() {
 
     // out1 receives both in1 and in2
     // out2 receives in2 only
-    let mut config = config::Config::empty();
+    let mut config = config::Config::builder();
     config.add_source(
         "in1",
         sources::socket::SocketConfig::make_tcp_config(in_addr1),
@@ -195,7 +195,7 @@ async fn merge_and_fork() {
     let mut output_lines1 = CountReceiver::receive_lines(out_addr1);
     let mut output_lines2 = CountReceiver::receive_lines(out_addr2);
 
-    let (topology, _crash) = start_topology(config, false).await;
+    let (topology, _crash) = start_topology(config.build(), false).await;
     // Wait for server to accept traffic
     wait_for_tcp(in_addr1).await;
     wait_for_tcp(in_addr2).await;
@@ -245,7 +245,7 @@ async fn reconnect() {
     let in_addr = next_addr();
     let out_addr = next_addr();
 
-    let mut config = config::Config::empty();
+    let mut config = config::Config::builder();
     config.add_source(
         "in",
         sources::socket::SocketConfig::make_tcp_config(in_addr),
@@ -258,7 +258,7 @@ async fn reconnect() {
 
     let output_lines = CountReceiver::receive_lines(out_addr);
 
-    let (topology, _crash) = start_topology(config, false).await;
+    let (topology, _crash) = start_topology(config.build(), false).await;
     // Wait for server to accept traffic
     wait_for_tcp(in_addr).await;
